removed R5 from VersionUtilities
if r5 is defined as a core package (in version utilities), then when the
validator loads the hl7.terminology package, it removes all code systems
loaded from the "core package" (i.e. r5) and replaces them with the ones
from the hl7.terminology package, which itself is for r4 and therefore
outdated. to prevent this behavior, r5 was removed again from the core
package list.